event/cnxk: fix QoS devargs parsing
[ upstream commit e21aa23bed1129cbe63ea621bfada4cc1ea55281 ]

Fix qos devargs parsing using incorrect datatype for the
structure elements.

Fixes: 38c2e3240ba8 ("event/cnxk: add option to control SSO HWGRP QoS")

+++ drivers/event/cnxk/cnxk_eventdev.c
@@ -482,7 +482,7 @@ static void
 parse_queue_param(char *value, void *opaque)
 {
 	struct cnxk_sso_qos queue_qos = {0};
-	uint8_t *val = (uint8_t *)&queue_qos;
+	uint16_t *val = (uint16_t *)&queue_qos;
 	struct cnxk_sso_evdev *dev = opaque;
 	char *tok = strtok(value, "-");
 	struct cnxk_sso_qos *old_ptr;
