+++ drivers/event/cnxk/cnxk_eventdev.h
@@ -74,9 +74,9 @@ typedef void (*cnxk_sso_hws_flush_t)(void *ws, uint8_t queue_id, uintptr_t base,
 
 struct cnxk_sso_qos {
 	uint16_t queue;
-	uint8_t xaq_prcnt;
-	uint8_t taq_prcnt;
-	uint8_t iaq_prcnt;
+	uint16_t xaq_prcnt;
+	uint16_t taq_prcnt;
+	uint16_t iaq_prcnt;
 };
 
 struct cnxk_sso_evdev {
